refactor: 게시글 상세 api 분리 및 사용하지 않는 함수와 기존 파일 삭제
