+++ src/features/community/hooks/usePostDetail.ts
@@ -1,4 +1,4 @@
-import { getPostDetail } from '@/api/community/posts';
+import { getPostDetail } from '@/src/features/community/apis/post';
 import { useQuery } from '@tanstack/react-query';
 
 export function usePostDetail(postId?: number) {
